msagent.js: Parse character info
msagent.js can now parse the character section of a ACS file fully.

This also adds a testbed thing. Later on it will display individual frames
and animations for debugging before the full character API is ready.
For now, it just has a file input for dumping ACS files into so we can debug
reading of the data.