msagent.js: Rewrite decompression to use WASM
Mostly so it actually works. I couldn't get a fully JS version working and I can't be bothered to when this seems to work.

Note that yes, the wasm IS checked in to the repository. This is just so clang isn't a direct build dependency, but will be needed if the decompression module needs to be updated.